Create LabInfrastructure project for holding common Lap code for use in integration and UI tests.
