+++ LabInfrastructure/KeyVaultConfiguration.cs
@@ -25,28 +25,38 @@
 //
 //------------------------------------------------------------------------------
 
-using System;
-using Xamarin.UITest;
-using Xamarin.UITest.Queries;
-
-namespace Test.Microsoft.Identity.Core.UIAutomation
+namespace LabInfrastructure
 {
-    /// <summary>
-    /// Initializes the app object that represents the main gateway to interact with the app on the device
-    /// </summary>
-	public class AppFactory
-	{
-        public static IApp StartApp(Platform platform, string targetApp)
-        {
-            switch (platform)
-            {
-                case Platform.Android:
-                    return ConfigureApp.Android.InstalledApp(targetApp).StartApp();
-                case Platform.iOS:
-                    return ConfigureApp.iOS.StartApp();
-                default:
-                    throw new PlatformNotSupportedException();
-            }
-        }
+    public class KeyVaultConfiguration
+    {
+        /// <summary>
+        /// The URL of the Key Vault instance.
+        /// </summary>
+        public string Url { get; set; }
+
+        /// <summary>
+        /// The authentication type to use to communicate with the Key Vault.
+        /// </summary>
+        public KeyVaultAuthenticationType AuthType { get; set; }
+
+        /// <summary>
+        /// The ID of the test harness client.
+        /// </summary>
+        /// <remarks>
+        /// This should be configured as having access to the Key Vault instance specified at <see cref="Url"/>.
+        /// </remarks>
+        public string ClientId { get; set; }
+
+        /// <summary>
+        /// The thumbprint of the <see cref="System.Security.Cryptography.X509Certificates.X509Certificate2"/> to use when
+        /// <see cref="AuthType"/> is <see cref="KeyVaultAuthenticationType.ClientCertificate"/>.
+        /// </summary>
+        public string CertThumbprint { get; set; }
+    }
+
+    public enum KeyVaultAuthenticationType
+    {
+        ClientCertificate,
+        UserCredential
     }
 }
