+++ LibsAndSamples.sln
@@ -102,7 +102,9 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AdalUAPTestApp", "adal\devA
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "XFormsApp.UWP", "adal\devApps\XFormsApp.UWP\XFormsApp.UWP.csproj", "{973ABA51-0B3A-4589-BF8D-19656C57D330}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Test.MSAL.NET.Integration", "msal\tests\Test.MSAL.NET.Integration\Test.MSAL.NET.Integration.csproj", "{F126A28F-C3FB-4729-BBCB-77EF9CEB76EF}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Test.MSAL.NET.Integration", "msal\tests\Test.MSAL.NET.Integration\Test.MSAL.NET.Integration.csproj", "{F126A28F-C3FB-4729-BBCB-77EF9CEB76EF}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LabInfrastructure", "LabInfrastructure\LabInfrastructure.csproj", "{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -2147,6 +2149,66 @@ Global
 		{F126A28F-C3FB-4729-BBCB-77EF9CEB76EF}.Release|x64.Build.0 = Release|Any CPU
 		{F126A28F-C3FB-4729-BBCB-77EF9CEB76EF}.Release|x86.ActiveCfg = Release|Any CPU
 		{F126A28F-C3FB-4729-BBCB-77EF9CEB76EF}.Release|x86.Build.0 = Release|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.Ad-Hoc|Any CPU.ActiveCfg = Debug|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.Ad-Hoc|Any CPU.Build.0 = Debug|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.Ad-Hoc|ARM.ActiveCfg = Debug|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.Ad-Hoc|ARM.Build.0 = Debug|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.Ad-Hoc|iPhone.ActiveCfg = Debug|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.Ad-Hoc|iPhone.Build.0 = Debug|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.Ad-Hoc|iPhoneSimulator.ActiveCfg = Debug|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.Ad-Hoc|iPhoneSimulator.Build.0 = Debug|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.Ad-Hoc|x64.ActiveCfg = Debug|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.Ad-Hoc|x64.Build.0 = Debug|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.Ad-Hoc|x86.ActiveCfg = Debug|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.Ad-Hoc|x86.Build.0 = Debug|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.AppStore|Any CPU.ActiveCfg = Debug|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.AppStore|Any CPU.Build.0 = Debug|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.AppStore|ARM.ActiveCfg = Debug|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.AppStore|ARM.Build.0 = Debug|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.AppStore|iPhone.ActiveCfg = Debug|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.AppStore|iPhone.Build.0 = Debug|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.AppStore|iPhoneSimulator.ActiveCfg = Debug|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.AppStore|iPhoneSimulator.Build.0 = Debug|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.AppStore|x64.ActiveCfg = Debug|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.AppStore|x64.Build.0 = Debug|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.AppStore|x86.ActiveCfg = Debug|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.AppStore|x86.Build.0 = Debug|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.Appveyor|Any CPU.ActiveCfg = Debug|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.Appveyor|Any CPU.Build.0 = Debug|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.Appveyor|ARM.ActiveCfg = Debug|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.Appveyor|ARM.Build.0 = Debug|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.Appveyor|iPhone.ActiveCfg = Debug|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.Appveyor|iPhone.Build.0 = Debug|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.Appveyor|iPhoneSimulator.ActiveCfg = Debug|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.Appveyor|iPhoneSimulator.Build.0 = Debug|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.Appveyor|x64.ActiveCfg = Debug|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.Appveyor|x64.Build.0 = Debug|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.Appveyor|x86.ActiveCfg = Debug|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.Appveyor|x86.Build.0 = Debug|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.Debug|ARM.ActiveCfg = Debug|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.Debug|ARM.Build.0 = Debug|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.Debug|iPhone.ActiveCfg = Debug|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.Debug|iPhone.Build.0 = Debug|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.Debug|iPhoneSimulator.ActiveCfg = Debug|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.Debug|iPhoneSimulator.Build.0 = Debug|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.Debug|x64.Build.0 = Debug|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.Debug|x86.Build.0 = Debug|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.Release|ARM.ActiveCfg = Release|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.Release|ARM.Build.0 = Release|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.Release|iPhone.ActiveCfg = Release|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.Release|iPhone.Build.0 = Release|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.Release|iPhoneSimulator.ActiveCfg = Release|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.Release|iPhoneSimulator.Build.0 = Release|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.Release|x64.ActiveCfg = Release|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.Release|x64.Build.0 = Release|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.Release|x86.ActiveCfg = Release|Any CPU
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -2198,6 +2260,7 @@ Global
 		{CF5DD35B-A52F-4C9C-B77C-04D15DF99D0F} = {54CB21B9-187C-4E63-9022-A9964886C56D}
 		{973ABA51-0B3A-4589-BF8D-19656C57D330} = {32A145D6-95A9-4232-B33E-AC4D9C392114}
 		{F126A28F-C3FB-4729-BBCB-77EF9CEB76EF} = {7F158B29-24C4-4FC0-BE4E-620573B43DEC}
+		{8AF6B089-8AEA-41C9-8890-C94B3E8876A0} = {9B0B5396-4D95-4C15-82ED-DC22B5A3123F}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {020399A9-DC27-4B82-9CAA-EF488665AC27}
