--- core/tests/Test.Microsoft.Identity.Core.UIAutomation/Infrastructure/UserType.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Test.Microsoft.Identity.Core.UIAutomation.infrastructure
-{
-    public enum UserType
-    {
-        Member = 0,
-        Guest = 1
-    }
-}
